Calculate precisions

+++ app/src/dialogs/orderwavesdialog/iorderwavesdialogfactory.h
@@ -23,7 +23,7 @@ public:
         IOrderWavesWidgetFactory* orderWavesWidgetFactory,
         IOrderBookThread*         orderBookThread,
         Stock*                    stock,
-        int                       precision,
+        qint8                     precision,
         QWidget*                  parent
     ) const = 0;
 };
