+++ app/src/dialogs/orderwavesdialog/iorderwavesdialogfactory_mock.h
@@ -26,7 +26,7 @@ public:
         (IOrderWavesWidgetFactory * orderWavesWidgetFactory,
          IOrderBookThread* orderBookThread,
          Stock*            stock,
-         int               precision,
+         qint8             precision,
          QWidget*          parent),
         (const, override)
     );
