+++ app/src/dialogs/orderwavesdialog/orderwavesdialogfactory.cpp
@@ -21,7 +21,7 @@ std::shared_ptr<IOrderWavesDialog> OrderWavesDialogFactory::newInstance(
     IOrderWavesWidgetFactory* orderWavesWidgetFactory,
     IOrderBookThread*         orderBookThread,
     Stock*                    stock,
-    int                       precision,
+    qint8                     precision,
     QWidget*                  parent
 ) const
 {
