+++ app/src/dialogs/orderwavesdialog/orderwavesdialogfactory.h
@@ -19,7 +19,7 @@ public:
         IOrderWavesWidgetFactory* orderWavesWidgetFactory,
         IOrderBookThread*         orderBookThread,
         Stock*                    stock,
-        int                       precision,
+        qint8                     precision,
         QWidget*                  parent
     ) const override;
 };
