+++ app/src/widgets/tableitems/actions/actionstableitemwidget.cpp
@@ -11,7 +11,7 @@ ActionsTableItemWidget::ActionsTableItemWidget(
     IOrderBookThread*         orderBookThread,
     IHttpClient*              httpClient,
     Stock*                    stock,
-    int                       precision,
+    qint8                     precision,
     QWidget*                  parent
 ) :
     IActionsTableItemWidget(parent),
